[Store] move remaining parallel tensor helpers to headers
Keep store_py.cpp focused on legacy/public wrappers by moving the remaining unified read/write helper logic into the parallel helper headers, while preserving the legacy TP read surface.
