+++ mooncake-integration/store/store_py.cpp
@@ -301,8 +301,7 @@ class MooncakeStorePyWrapper {
         return store_->health_check();
     }
 
-    std::string get_tp_key_name(const std::string &base_key,
-                                int rank) const {
+    std::string get_tp_key_name(const std::string &base_key, int rank) const {
         return base_key + "_tp_" + std::to_string(rank);
     }
 
@@ -631,20 +630,6 @@ class MooncakeStorePyWrapper {
                                        tp_size, split_dim);
     }
 
-    int put_tensor_parallelism_tp_impl(
-        const std::string &key, pybind11::object tensor,
-        const ReplicateConfig &config, int tp_rank, int tp_size, int split_dim,
-        const std::vector<ParallelAxisSpec> &axes) {
-        return execute_tp_tensor_write_impl(
-            key, tensor, config, tp_size, split_dim, axes,
-            "Failed to put tensor with tp",
-            [this](const std::string &shard_key, const PyTensorInfo &info,
-                   const ReplicateConfig &write_config) {
-                return put_tensor_info_impl(shard_key, info, write_config);
-            });
-    }
-
-
     std::vector<int> batch_put_tensor_infos_impl(
         const std::vector<std::string> &keys,
         const std::vector<PyTensorInfo> &infos,
@@ -844,185 +829,8 @@ class MooncakeStorePyWrapper {
         return store_->batch_put_from(keys, buffers, sizes, ReplicateConfig{});
     }
 
-    pybind11::object get_tensor_with_writer_shard_full(
-        const std::string &key, const std::string &context) {
-        if (!is_client_initialized()) {
-            LOG(ERROR) << "Client is not initialized";
-            return pybind11::none();
-        }
-        if (use_dummy_client_) {
-            LOG(ERROR) << context << ": dummy client is not supported";
-            return pybind11::none();
-        }
-
-        auto reconstruction = load_writer_shard_reconstruction(key, context);
-        if (!reconstruction.has_value()) {
-            return get_tensor(key);
-        }
-
-        size_t total_tensor_numel = 1;
-        for (auto dim : reconstruction->global_shape) {
-            total_tensor_numel *= static_cast<size_t>(dim);
-        }
-
-        size_t element_size = 0;
-        for (const auto &source : reconstruction->sources) {
-            if (source.metadata.data_bytes == 0) {
-                continue;
-            }
-            int64_t shard_numel = 1;
-            const auto local_shape =
-                TensorShapeToVector(source.metadata.metadata.layout.local_shape,
-                                    source.metadata.metadata.header.ndim);
-            for (auto dim : local_shape) {
-                shard_numel *= dim;
-            }
-            if (shard_numel <= 0 ||
-                source.metadata.data_bytes % static_cast<size_t>(shard_numel) !=
-                    0) {
-                LOG(ERROR) << context
-                           << ": invalid writer shard tensor byte size";
-                return py::none();
-            }
-            element_size =
-                source.metadata.data_bytes / static_cast<size_t>(shard_numel);
-            break;
-        }
-
-        const size_t total_length =
-            sizeof(TensorMetadata) + total_tensor_numel * element_size;
-        char *owned_buffer = new char[total_length];
-        if (store_->register_buffer(owned_buffer, total_length) != 0) {
-            LOG(ERROR) << context
-                       << ": failed to register reconstruction buffer";
-            delete[] owned_buffer;
-            return py::none();
-        }
-
-        auto plan = build_full_tensor_into_plan_from_sources(
-            reinterpret_cast<uintptr_t>(owned_buffer), total_length,
-            reconstruction->sources, reconstruction->global_shape,
-            reconstruction->split_dim, reconstruction->dtype, context,
-            reconstruction->allow_empty_fragments);
-        if (!plan.has_value()) {
-            store_->unregister_buffer(owned_buffer);
-            delete[] owned_buffer;
-            return py::none();
-        }
-
-        auto success = execute_tensor_into_plan_transfers({*plan});
-        if (success.empty() || !success[0]) {
-            store_->unregister_buffer(owned_buffer);
-            delete[] owned_buffer;
-            return py::none();
-        }
-
-        return buffer_to_tensor(
-            new BufferHandle(owned_buffer, total_length,
-                             [this, owned_buffer]() {
-                                 store_->unregister_buffer(owned_buffer);
-                                 delete[] owned_buffer;
-                             }),
-            nullptr, 0);
-    }
-
-    pybind11::object get_tensor_with_tp_full(
-        const std::string &key, int tp_rank, int tp_size, int split_dim,
-        const std::string &context,
-        const std::optional<TensorParallelismSpec> &parallelism =
-            std::nullopt) {
-        if (!is_client_initialized()) {
-            LOG(ERROR) << "Client is not initialized";
-            return pybind11::none();
-        }
-        if (use_dummy_client_) {
-            LOG(ERROR) << context << ": dummy client is not supported";
-            return pybind11::none();
-        }
-        ParallelAxisSpec axis{
-            .kind = "tp",
-            .rank = tp_rank,
-            .size = tp_size,
-            .split_dim = split_dim,
-            .expert_id = std::nullopt,
-            .stage_id = std::nullopt,
-        };
-
-        auto reconstruction =
-            parallelism.has_value()
-                ? load_parallelism_full_reconstruction_sources(
-                      key, *parallelism, context)
-                : load_tp_full_reconstruction_sources(key, axis, context);
-        if (!reconstruction.has_value()) {
-            return pybind11::none();
-        }
-
-        const size_t total_tensor_numel =
-            std::accumulate(reconstruction->global_shape.begin(),
-                            reconstruction->global_shape.end(),
-                            static_cast<size_t>(1), std::multiplies<size_t>());
-        size_t element_size = 0;
-        for (const auto &source : reconstruction->sources) {
-            if (source.metadata.data_bytes == 0) {
-                continue;
-            }
-            int64_t shard_numel = 1;
-            const auto local_shape =
-                TensorShapeToVector(source.metadata.metadata.layout.local_shape,
-                                    source.metadata.metadata.header.ndim);
-            for (auto dim : local_shape) {
-                shard_numel *= dim;
-            }
-            if (shard_numel <= 0 ||
-                source.metadata.data_bytes % static_cast<size_t>(shard_numel) !=
-                    0) {
-                LOG(ERROR) << context << ": invalid shard tensor byte size";
-                return pybind11::none();
-            }
-            element_size =
-                source.metadata.data_bytes / static_cast<size_t>(shard_numel);
-            break;
-        }
-        const size_t total_length =
-            sizeof(TensorMetadata) + total_tensor_numel * element_size;
-
-        char *owned_buffer = new char[total_length];
-        if (store_->register_buffer(owned_buffer, total_length) != 0) {
-            LOG(ERROR) << context
-                       << ": failed to register reconstruction buffer";
-            delete[] owned_buffer;
-            return pybind11::none();
-        }
-
-        auto plan = build_full_tensor_into_plan_from_sources(
-            reinterpret_cast<uintptr_t>(owned_buffer), total_length,
-            reconstruction->sources, reconstruction->global_shape,
-            reconstruction->split_dim, reconstruction->dtype, context,
-            reconstruction->allow_empty_fragments);
-        if (!plan.has_value()) {
-            store_->unregister_buffer(owned_buffer);
-            delete[] owned_buffer;
-            return pybind11::none();
-        }
-
-        auto success = execute_tensor_into_plan_transfers({*plan});
-        if (success.empty() || !success[0]) {
-            store_->unregister_buffer(owned_buffer);
-            delete[] owned_buffer;
-            return py::none();
-        }
-
-        return buffer_to_tensor(
-            new BufferHandle(owned_buffer, total_length,
-                             [this, owned_buffer]() {
-                                 store_->unregister_buffer(owned_buffer);
-                                 delete[] owned_buffer;
-                             }),
-            nullptr, 0);
-    }
-
     int put_tensor_info_impl(const std::string &key, const PyTensorInfo &info,
-                        const ReplicateConfig &config) {
+                             const ReplicateConfig &config) {
         if (!info.valid()) return to_py_ret(ErrorCode::INVALID_PARAMS);
 
         std::vector<std::span<const char>> values;
@@ -1037,753 +845,6 @@ class MooncakeStorePyWrapper {
             LOG(ERROR) << "put_parts failed for key " << key << " with code "
                        << ret;
         return ret;
-    }
-
-    template <typename ManifestWriteFn>
-    int write_manifest_impl(const std::string &key,
-                            const WriterShardManifest &manifest,
-                            const char *operation_name,
-                            ManifestWriteFn &&write_manifest) {
-        std::span<const char> bytes(reinterpret_cast<const char *>(&manifest),
-                                    sizeof(WriterShardManifest));
-        py::gil_scoped_release release_gil;
-        int ret = write_manifest(bytes);
-        if (ret != 0) {
-            LOG(ERROR) << operation_name << " manifest failed for key " << key
-                       << " with code " << ret;
-        }
-        return ret;
-    }
-
-    template <typename BatchWriteFromFn>
-    std::vector<int> batch_write_tensor_impl(
-        const std::vector<std::string> &keys,
-        const std::vector<PyTensorInfo> &infos, const ReplicateConfig &config,
-        const char *operation_name, BatchWriteFromFn &&batch_write_from) {
-        std::vector<int> results(keys.size(), 0);
-
-        {
-            py::gil_scoped_release release_gil;
-
-            std::vector<std::string> valid_keys;
-            std::vector<void *> buffer_ptrs;
-            std::vector<size_t> buffer_sizes;
-            std::vector<size_t> original_indices;
-            std::vector<std::unique_ptr<BufferHandle>> temp_allocations;
-
-            for (size_t i = 0; i < infos.size(); ++i) {
-                if (!infos[i].valid()) {
-                    results[i] = to_py_ret(ErrorCode::INVALID_PARAMS);
-                    continue;
-                }
-
-                size_t total_size =
-                    infos[i].metadata.header.data_offset + infos[i].tensor_size;
-                auto alloc_result =
-                    store_->client_buffer_allocator_->allocate(total_size);
-
-                if (!alloc_result) {
-                    LOG(ERROR) << "Failed to allocate buffer for "
-                               << operation_name << " key: " << keys[i];
-                    results[i] = to_py_ret(ErrorCode::INVALID_PARAMS);
-                    continue;
-                }
-
-                char *dst = static_cast<char *>(alloc_result->ptr());
-                std::memcpy(dst, &infos[i].metadata,
-                            infos[i].metadata.header.data_offset);
-                std::memcpy(dst + infos[i].metadata.header.data_offset,
-                            reinterpret_cast<void *>(infos[i].data_ptr),
-                            infos[i].tensor_size);
-
-                valid_keys.push_back(keys[i]);
-                buffer_ptrs.push_back(alloc_result->ptr());
-                buffer_sizes.push_back(total_size);
-                original_indices.push_back(i);
-                temp_allocations.push_back(
-                    std::make_unique<BufferHandle>(std::move(*alloc_result)));
-            }
-
-            if (!valid_keys.empty()) {
-                std::vector<int> op_results =
-                    batch_write_from(valid_keys, buffer_ptrs, buffer_sizes);
-                for (size_t i = 0; i < op_results.size(); ++i) {
-                    results[original_indices[i]] = op_results[i];
-                }
-            }
-        }
-
-        return results;
-    }
-
-    bool ensure_tensor_write_supported(const char *operation_name) const {
-        if (!is_client_initialized() || use_dummy_client_) {
-            LOG(ERROR) << operation_name
-                       << ": client not initialized or dummy client not "
-                          "supported for tensors";
-            return false;
-        }
-        return true;
-    }
-
-    template <typename WriteFn>
-    int execute_single_tensor_write(const char *operation_name,
-                                    const ReplicateConfig &config,
-                                    WriteFn &&write_fn) {
-        if (!ensure_tensor_write_supported(operation_name)) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-        int validate_result = validate_replicate_config(config);
-        if (validate_result) {
-            return validate_result;
-        }
-        return write_fn();
-    }
-
-    template <typename WriteFn>
-    std::vector<int> execute_batch_tensor_write(
-        const char *operation_name, const char *size_error_context,
-        const std::vector<std::string> &keys, size_t value_count,
-        const ReplicateConfig &config, WriteFn &&write_fn) {
-        if (!ensure_tensor_write_supported(operation_name)) {
-            return std::vector<int>(keys.size(),
-                                    to_py_ret(ErrorCode::INVALID_PARAMS));
-        }
-        if (keys.size() != value_count || keys.empty()) {
-            if (!keys.empty()) {
-                LOG(ERROR) << size_error_context;
-            }
-            return std::vector<int>(keys.size(),
-                                    to_py_ret(ErrorCode::INVALID_PARAMS));
-        }
-        int validate_result = validate_replicate_config(config);
-        if (validate_result) {
-            return std::vector<int>(keys.size(), validate_result);
-        }
-        return write_fn();
-    }
-
-    bool validate_tensor_object_buffers(
-        const std::vector<std::string> &keys,
-        const std::vector<uintptr_t> &buffer_ptrs,
-        const std::vector<size_t> &sizes, const char *size_error_context,
-        const char *buffer_error_context) {
-        if (keys.size() != buffer_ptrs.size() || keys.size() != sizes.size()) {
-            LOG(ERROR) << size_error_context;
-            return false;
-        }
-        for (size_t i = 0; i < sizes.size(); ++i) {
-            if (!is_valid_tensor_object_buffer(
-                    buffer_ptrs[i], sizes[i],
-                    std::string(buffer_error_context) + " at index " +
-                        std::to_string(i))) {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    template <typename WriteFromFn>
-    int execute_single_tensor_write_from(const char *operation_name,
-                                         const std::string &key,
-                                         uintptr_t buffer_ptr, size_t size,
-                                         const ReplicateConfig &config,
-                                         WriteFromFn &&write_from_fn) {
-        if (!ensure_tensor_write_supported(operation_name)) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-        if (!is_valid_tensor_object_buffer(buffer_ptr, size, operation_name)) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-        int validate_result = validate_replicate_config(config);
-        if (validate_result) {
-            return validate_result;
-        }
-        return write_from_fn(key, reinterpret_cast<void *>(buffer_ptr), size,
-                             config);
-    }
-
-    template <typename BatchWriteFromFn>
-    std::vector<int> execute_batch_tensor_write_from(
-        const char *operation_name, const char *size_error_context,
-        const std::vector<std::string> &keys,
-        const std::vector<uintptr_t> &buffer_ptrs,
-        const std::vector<size_t> &sizes, const ReplicateConfig &config,
-        BatchWriteFromFn &&batch_write_from_fn) {
-        if (!ensure_tensor_write_supported(operation_name)) {
-            return std::vector<int>(keys.size(),
-                                    to_py_ret(ErrorCode::INVALID_PARAMS));
-        }
-        if (keys.empty()) {
-            return std::vector<int>();
-        }
-        if (!validate_tensor_object_buffers(keys, buffer_ptrs, sizes,
-                                            size_error_context,
-                                            "tensor object buffer")) {
-            return std::vector<int>(keys.size(),
-                                    to_py_ret(ErrorCode::INVALID_PARAMS));
-        }
-        int validate_result = validate_replicate_config(config);
-        if (validate_result) {
-            return std::vector<int>(keys.size(), validate_result);
-        }
-        std::vector<void *> buffers;
-        buffers.reserve(buffer_ptrs.size());
-        for (uintptr_t ptr : buffer_ptrs) {
-            buffers.push_back(reinterpret_cast<void *>(ptr));
-        }
-        return batch_write_from_fn(keys, buffers, sizes, config);
-    }
-
-    template <typename DirectWriteFn, typename ParallelismWriteFn,
-              typename WriterPartitionWriteFn>
-    std::vector<int> execute_batch_parallelism_write_requests(
-        const std::vector<std::string> &keys, size_t value_count,
-        const py::object &parallelisms, const py::object &writer_partitions,
-        const char *error_context, DirectWriteFn &&direct_write,
-        ParallelismWriteFn &&parallelism_write,
-        WriterPartitionWriteFn &&writer_partition_write) {
-        if (!parallelisms.is_none() && !writer_partitions.is_none()) {
-            LOG(ERROR)
-                << error_context
-                << ": writer_partitions cannot be combined with parallelisms";
-            return std::vector<int>(keys.size(),
-                                    to_py_ret(ErrorCode::INVALID_PARAMS));
-        }
-        if (value_count != keys.size()) {
-            LOG(ERROR) << error_context
-                       << ": values and keys must have the same length";
-            return std::vector<int>(keys.size(),
-                                    to_py_ret(ErrorCode::INVALID_PARAMS));
-        }
-        if (parallelisms.is_none() && writer_partitions.is_none()) {
-            return direct_write();
-        }
-
-        std::vector<int> results(keys.size(),
-                                 to_py_ret(ErrorCode::INVALID_PARAMS));
-        if (!parallelisms.is_none()) {
-            auto parallelism_list = validate_batch_parallelism_list(
-                parallelisms, keys.size(), error_context);
-            if (!parallelism_list.has_value()) {
-                return std::vector<int>(keys.size(),
-                                        to_py_ret(ErrorCode::INVALID_PARAMS));
-            }
-            for (size_t i = 0; i < keys.size(); ++i) {
-                results[i] = parallelism_write(i, (*parallelism_list)[i]);
-            }
-            return results;
-        }
-
-        auto writer_partition_list = validate_batch_writer_partition_list(
-            writer_partitions, keys.size(), error_context);
-        if (!writer_partition_list.has_value()) {
-            return std::vector<int>(keys.size(),
-                                    to_py_ret(ErrorCode::INVALID_PARAMS));
-        }
-        for (size_t i = 0; i < keys.size(); ++i) {
-            results[i] = writer_partition_write(i, (*writer_partition_list)[i]);
-        }
-        return results;
-    }
-
-    template <typename DirectWriteFn, typename WriterShardWriteFn,
-              typename LegacyTpWriteFn, typename TpParallelWriteFn,
-              typename GenericShardWriteFn>
-    int execute_parallelism_tensor_write_route(
-        const std::string &key, pybind11::object tensor,
-        const ResolvedParallelismWriteRequest &request,
-        const ReplicateConfig &config, const char *error_context,
-        DirectWriteFn &&direct_write, WriterShardWriteFn &&writer_shard_write,
-        LegacyTpWriteFn &&legacy_tp_write,
-        TpParallelWriteFn &&tp_parallel_write,
-        GenericShardWriteFn &&generic_shard_write) {
-        switch (request.route) {
-            case ParallelismWriteStorageRoute::DIRECT_FULL_OBJECT:
-                if (!ensure_tensor_write_supported(error_context)) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                if (int validate_result = validate_replicate_config(config);
-                    validate_result) {
-                    return validate_result;
-                }
-                return direct_write(key, tensor, config);
-            case ParallelismWriteStorageRoute::WRITER_PARTITION_SHARD:
-                if (!request.writer_partition.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                return writer_shard_write(key, tensor,
-                                          *request.writer_partition, config);
-            case ParallelismWriteStorageRoute::LEGACY_SINGLE_TP:
-            case ParallelismWriteStorageRoute::TP_SHARDED_PARALLELISM:
-            case ParallelismWriteStorageRoute::GENERIC_PARALLELISM_SHARD: {
-                if (!request.parallelism.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                switch (request.route) {
-                    case ParallelismWriteStorageRoute::LEGACY_SINGLE_TP: {
-                        const auto &axis = request.parallelism->axes[0];
-                        return legacy_tp_write(key, tensor, config, axis.rank,
-                                               axis.size,
-                                               axis.split_dim.value_or(0));
-                    }
-                    case ParallelismWriteStorageRoute::TP_SHARDED_PARALLELISM: {
-                        auto tp_axis_index =
-                            find_tp_axis_index(request.parallelism->axes);
-                        if (!tp_axis_index.has_value()) {
-                            return to_py_ret(ErrorCode::INVALID_PARAMS);
-                        }
-                        const auto &axis =
-                            request.parallelism->axes[*tp_axis_index];
-                        return tp_parallel_write(key, tensor, config, axis.rank,
-                                                 axis.size,
-                                                 axis.split_dim.value_or(0),
-                                                 request.parallelism->axes);
-                    }
-                    case ParallelismWriteStorageRoute::
-                        GENERIC_PARALLELISM_SHARD:
-                        return generic_shard_write(
-                            key, tensor, *request.parallelism, config);
-                    default:
-                        break;
-                }
-                break;
-            }
-        }
-        return to_py_ret(ErrorCode::INVALID_PARAMS);
-    }
-
-    template <typename WritePartsFn>
-    int execute_tensor_parts_write(
-        const std::string &key,
-        const std::vector<std::span<const char>> &values,
-        const ReplicateConfig &config, const TensorWriteStoreOps &ops,
-        WritePartsFn &&write_parts) {
-        py::gil_scoped_release release_gil;
-        int ret = write_parts(key, values, config);
-        if (ret != 0) {
-            LOG(ERROR) << ops.parts_operation_name << " failed for key " << key
-                       << " with code " << ret;
-        }
-        return ret;
-    }
-
-    template <typename DirectWriteFn>
-    int execute_parallelism_direct_tensor_write_from(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const ReplicateConfig &config, const char *operation_name,
-        DirectWriteFn &&direct_write) {
-        if (is_default_replicate_config(config)) {
-            return direct_write(key, buffer_ptr, size);
-        }
-        return execute_single_tensor_write_from(
-            operation_name, key, buffer_ptr, size, config,
-            [&](const std::string &write_key, void *buffer, size_t write_size,
-                const ReplicateConfig &write_config) {
-                return direct_write(write_key,
-                                    reinterpret_cast<uintptr_t>(buffer),
-                                    write_size, write_config);
-            });
-    }
-
-    pybind11::object decode_tensor_object_buffer(uintptr_t buffer_ptr,
-                                                 size_t size,
-                                                 const char *operation_name) {
-        if (!is_valid_tensor_object_buffer(buffer_ptr, size, operation_name)) {
-            return py::none();
-        }
-        pybind11::object tensor =
-            buffer_to_tensor(NULL, reinterpret_cast<char *>(buffer_ptr),
-                             static_cast<int64_t>(size));
-        if (tensor.is_none()) {
-            LOG(ERROR) << "Failed to decode tensor buffer for "
-                       << operation_name;
-        }
-        return tensor;
-    }
-
-    std::string get_tp_write_shard_key(
-        const std::string &base_key, int rank,
-        const std::vector<ParallelAxisSpec> &axes) {
-        if (axes.empty()) {
-            return get_tp_key_name(base_key, rank);
-        }
-        auto shard_axes = axes;
-        auto tp_axis_index = find_tp_axis_index(shard_axes);
-        if (!tp_axis_index.has_value()) {
-            return std::string();
-        }
-        shard_axes[*tp_axis_index].rank = rank;
-        return get_parallelism_key_name(base_key,
-                                        TensorParallelismSpec{shard_axes});
-    }
-
-    template <typename WriteShardFn>
-    int execute_tp_tensor_write_impl(const std::string &key,
-                                     pybind11::object tensor,
-                                     const ReplicateConfig &config, int tp_size,
-                                     int split_dim,
-                                     const std::vector<ParallelAxisSpec> &axes,
-                                     const char *error_context,
-                                     WriteShardFn &&write_shard) {
-        try {
-            auto shard_infos = build_tp_shard_infos(
-                tensor, tp_size, split_dim,
-                [&](int rank) {
-                    return get_tp_write_shard_key(key, rank, axes);
-                },
-                axes);
-            if (!shard_infos.has_value()) {
-                LOG(ERROR) << error_context;
-                return to_py_ret(ErrorCode::INVALID_PARAMS);
-            }
-
-            for (int rank = 0; rank < tp_size; ++rank) {
-                std::string shard_key = get_tp_write_shard_key(key, rank, axes);
-                if (shard_key.empty()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                int ret = write_shard(shard_key, (*shard_infos)[rank], config);
-                if (ret != 0) return ret;
-            }
-            return 0;
-        } catch (const std::exception &e) {
-            LOG(ERROR) << error_context << ": " << e.what();
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-    }
-
-    template <typename WritePartsFn>
-    int execute_tp_tensor_write_from_impl(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const ReplicateConfig &config, int tp_size, int split_dim,
-        const std::vector<ParallelAxisSpec> &axes, const char *error_context,
-        WritePartsFn &&write_parts) {
-        auto parsed = parse_tensor_metadata_from_raw_buffer(buffer_ptr, size,
-                                                            error_context);
-        if (!parsed.has_value()) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-
-        for (int rank = 0; rank < tp_size; ++rank) {
-            std::string shard_key = get_tp_write_shard_key(key, rank, axes);
-            if (shard_key.empty()) {
-                return to_py_ret(ErrorCode::INVALID_PARAMS);
-            }
-
-            auto plan = build_raw_tensor_shard_write_plan(
-                *parsed, split_dim, rank, tp_size, error_context);
-            if (!plan.has_value()) {
-                return to_py_ret(ErrorCode::INVALID_PARAMS);
-            }
-            const auto global_shape =
-                TensorShapeToVector(parsed->metadata.layout.global_shape,
-                                    parsed->metadata.header.ndim);
-            const auto local_shape = TensorShapeToVector(
-                plan->metadata.layout.local_shape, plan->metadata.header.ndim);
-            if (axes.empty()) {
-                ParallelAxisSpec axis_spec{
-                    "tp", rank, tp_size, split_dim, std::nullopt, std::nullopt};
-                auto metadata = build_shard_metadata_from_shapes(
-                    parsed->metadata.header.dtype, global_shape, local_shape,
-                    {axis_spec}, plan->metadata.header.data_bytes);
-                if (!metadata.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                plan->metadata = *metadata;
-            } else {
-                auto shard_axes = axes;
-                auto tp_axis_index = find_tp_axis_index(shard_axes);
-                if (!tp_axis_index.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                shard_axes[*tp_axis_index].rank = rank;
-                const auto local_shape =
-                    TensorShapeToVector(plan->metadata.layout.local_shape,
-                                        plan->metadata.header.ndim);
-                auto metadata = build_shard_metadata_from_shapes(
-                    parsed->metadata.header.dtype, global_shape, local_shape,
-                    shard_axes, plan->metadata.header.data_bytes);
-                if (!metadata.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                plan->metadata = *metadata;
-            }
-
-            std::vector<std::span<const char>> values;
-            values.emplace_back(reinterpret_cast<const char *>(&plan->metadata),
-                                plan->metadata.header.data_offset);
-            for (const auto &[src_offset, part_size] : plan->data_ranges) {
-                values.emplace_back(
-                    reinterpret_cast<const char *>(buffer_ptr + src_offset),
-                    part_size);
-            }
-            int ret = write_parts(shard_key, values, config);
-            if (ret != 0) {
-                return ret;
-            }
-        }
-        return 0;
-    }
-
-    template <typename BatchWriteFn>
-    std::vector<int> batch_execute_tp_tensor_write_impl(
-        const std::vector<std::string> &base_keys,
-        const pybind11::list &tensors_list, const ReplicateConfig &config,
-        int tp_size, int split_dim, const char *error_context,
-        BatchWriteFn &&batch_write) {
-        std::vector<std::string> all_chunk_keys;
-        std::vector<PyTensorInfo> all_chunk_infos;
-        std::vector<size_t> processed_indices;
-        std::vector<int> final_results(base_keys.size(),
-                                       to_py_ret(ErrorCode::INVALID_PARAMS));
-        try {
-            for (size_t i = 0; i < base_keys.size(); ++i) {
-                py::object tensor = tensors_list[i];
-                if (tensor.is_none()) {
-                    final_results[i] = to_py_ret(ErrorCode::INVALID_PARAMS);
-                    continue;
-                }
-
-                auto shard_infos = build_tp_shard_infos(
-                    tensor, tp_size, split_dim, [&](int rank) {
-                        return get_tp_key_name(base_keys[i], rank);
-                    });
-                if (!shard_infos.has_value()) {
-                    final_results[i] = to_py_ret(ErrorCode::INVALID_PARAMS);
-                    continue;
-                }
-
-                processed_indices.push_back(i);
-                for (int rank = 0; rank < tp_size; ++rank) {
-                    all_chunk_keys.push_back(
-                        get_tp_key_name(base_keys[i], rank));
-                    all_chunk_infos.push_back((*shard_infos)[rank]);
-                }
-            }
-
-            if (all_chunk_keys.empty()) return final_results;
-
-            std::vector<int> chunk_results =
-                batch_write(all_chunk_keys, all_chunk_infos, config);
-            for (size_t i = 0; i < processed_indices.size(); ++i) {
-                size_t original_idx = processed_indices[i];
-                bool all_ok = true;
-                for (int j = 0; j < tp_size; ++j) {
-                    int res = chunk_results[i * tp_size + j];
-                    if (res != 0) {
-                        final_results[original_idx] = res;
-                        all_ok = false;
-                        break;
-                    }
-                }
-                if (all_ok) {
-                    final_results[original_idx] = 0;
-                }
-            }
-        } catch (const std::exception &e) {
-            LOG(ERROR) << error_context << ": " << e.what();
-        }
-
-        return final_results;
-    }
-
-    template <typename BatchTensorWriteFn>
-    std::vector<int> batch_decode_tensor_buffers_and_write(
-        const std::vector<std::string> &keys,
-        const std::vector<uintptr_t> &buffer_ptrs,
-        const std::vector<size_t> &sizes, const char *decode_error_context,
-        BatchTensorWriteFn &&batch_tensor_write_fn) {
-        py::list tensors_list;
-        std::vector<size_t> processed_indices;
-        std::vector<int> final_results(keys.size(), 0);
-
-        for (size_t i = 0; i < keys.size(); ++i) {
-            if (!is_valid_tensor_object_buffer(
-                    buffer_ptrs[i], sizes[i],
-                    std::string("tensor object buffer at index ") +
-                        std::to_string(i))) {
-                final_results[i] = to_py_ret(ErrorCode::INVALID_PARAMS);
-                continue;
-            }
-
-            py::object tensor =
-                buffer_to_tensor(NULL, reinterpret_cast<char *>(buffer_ptrs[i]),
-                                 static_cast<int64_t>(sizes[i]));
-            if (tensor.is_none()) {
-                LOG(ERROR) << decode_error_context << " at index " << i;
-                final_results[i] = to_py_ret(ErrorCode::INVALID_PARAMS);
-                continue;
-            }
-            tensors_list.append(tensor);
-            processed_indices.push_back(i);
-        }
-
-        if (processed_indices.empty()) {
-            return final_results;
-        }
-
-        std::vector<std::string> valid_keys;
-        valid_keys.reserve(processed_indices.size());
-        for (size_t idx : processed_indices) {
-            valid_keys.push_back(keys[idx]);
-        }
-
-        std::vector<int> op_results =
-            batch_tensor_write_fn(valid_keys, tensors_list);
-        for (size_t i = 0; i < processed_indices.size(); ++i) {
-            final_results[processed_indices[i]] = op_results[i];
-        }
-        return final_results;
-    }
-
-    template <typename WriteShardFn>
-    int execute_direct_parallelism_shard_write_from(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const TensorParallelismSpec &parallelism, const ReplicateConfig &config,
-        const char *error_context, const TensorWriteStoreOps &ops,
-        WriteShardFn &&write_shard) {
-        auto parsed = parse_tensor_metadata_from_raw_buffer(buffer_ptr, size,
-                                                            error_context);
-        if (!parsed.has_value()) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-
-        auto metadata = build_shard_metadata_from_shapes(
-            parsed->metadata.header.dtype,
-            TensorShapeToVector(parsed->metadata.layout.global_shape,
-                                parsed->metadata.header.ndim),
-            TensorShapeToVector(parsed->metadata.layout.local_shape,
-                                parsed->metadata.header.ndim),
-            parallelism.axes, parsed->data_bytes);
-        if (!metadata.has_value()) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-
-        std::string shard_key = get_parallelism_key_name(key, parallelism);
-        std::vector<std::span<const char>> values;
-        values.emplace_back(reinterpret_cast<const char *>(&*metadata),
-                            metadata->header.data_offset);
-        values.emplace_back(
-            reinterpret_cast<const char *>(buffer_ptr + parsed->data_offset),
-            parsed->data_bytes);
-        return execute_tensor_parts_write(
-            shard_key, values, config, ops,
-            std::forward<WriteShardFn>(write_shard));
-    }
-
-    template <typename WritePartsFn, typename WriteManifestFn>
-    int execute_writer_partition_shard_write_from(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const WriterPartitionSpec &writer, const ReplicateConfig &config,
-        const char *error_context, const TensorWriteStoreOps &ops,
-        WritePartsFn &&write_parts, WriteManifestFn &&write_manifest) {
-        auto parsed = parse_tensor_metadata_from_raw_buffer(buffer_ptr, size,
-                                                            error_context);
-        if (!parsed.has_value()) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-
-        auto plan = build_raw_tensor_shard_write_plan(
-            *parsed, writer.split_dim, writer.rank, writer.size, error_context);
-        if (!plan.has_value()) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-
-        ParallelAxisSpec axis_spec{"tp",         writer.rank,
-                                   writer.size,  writer.split_dim,
-                                   std::nullopt, std::nullopt};
-        auto metadata = build_shard_metadata_from_shapes(
-            parsed->metadata.header.dtype,
-            TensorShapeToVector(parsed->metadata.layout.global_shape,
-                                parsed->metadata.header.ndim),
-            TensorShapeToVector(plan->metadata.layout.local_shape,
-                                plan->metadata.header.ndim),
-            {axis_spec}, plan->metadata.header.data_bytes);
-        if (!metadata.has_value()) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-        plan->metadata = *metadata;
-
-        std::vector<std::span<const char>> values;
-        values.emplace_back(reinterpret_cast<const char *>(&plan->metadata),
-                            plan->metadata.header.data_offset);
-        for (const auto &[src_offset, part_size] : plan->data_ranges) {
-            values.emplace_back(
-                reinterpret_cast<const char *>(buffer_ptr + src_offset),
-                part_size);
-        }
-
-        int ret = execute_tensor_parts_write(
-            get_writer_shard_key_name(key, writer), values, config, ops,
-            std::forward<WritePartsFn>(write_parts));
-        if (ret != 0) {
-            return ret;
-        }
-        return write_manifest(
-            get_writer_manifest_key_name(key),
-            build_writer_shard_manifest_from_shape(
-                TensorShapeToVector(parsed->metadata.layout.global_shape,
-                                    parsed->metadata.header.ndim),
-                parsed->metadata.header.dtype, writer),
-            config);
-    }
-
-    template <typename DirectFromWriteFn, typename WriterShardFromWriteFn,
-              typename TpShardFromWriteFn, typename GenericShardFromWriteFn,
-              typename TensorRouteWriteFn>
-    int execute_parallelism_tensor_write_from_route(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const ResolvedParallelismWriteRequest &request,
-        const ReplicateConfig &config, const char *error_context,
-        DirectFromWriteFn &&direct_from_write,
-        WriterShardFromWriteFn &&writer_shard_from_write,
-        TpShardFromWriteFn &&tp_shard_from_write,
-        GenericShardFromWriteFn &&generic_shard_from_write,
-        TensorRouteWriteFn &&tensor_route_write) {
-        switch (request.route) {
-            case ParallelismWriteStorageRoute::DIRECT_FULL_OBJECT:
-                return direct_from_write(key, buffer_ptr, size, config);
-            case ParallelismWriteStorageRoute::WRITER_PARTITION_SHARD:
-                if (!request.writer_partition.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                return writer_shard_from_write(
-                    key, buffer_ptr, size, *request.writer_partition, config);
-            case ParallelismWriteStorageRoute::LEGACY_SINGLE_TP:
-            case ParallelismWriteStorageRoute::TP_SHARDED_PARALLELISM: {
-                if (!request.parallelism.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                auto tp_axis_index =
-                    find_tp_axis_index(request.parallelism->axes);
-                if (!tp_axis_index.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                const auto &axis = request.parallelism->axes[*tp_axis_index];
-                const auto axes =
-                    request.route ==
-                            ParallelismWriteStorageRoute::LEGACY_SINGLE_TP
-                        ? std::vector<ParallelAxisSpec>{}
-                        : request.parallelism->axes;
-                return tp_shard_from_write(key, buffer_ptr, size, axis.size,
-                                           axis.split_dim.value_or(0), axes,
-                                           config);
-            }
-            case ParallelismWriteStorageRoute::GENERIC_PARALLELISM_SHARD:
-                if (!request.parallelism.has_value()) {
-                    return to_py_ret(ErrorCode::INVALID_PARAMS);
-                }
-                return generic_shard_from_write(key, buffer_ptr, size,
-                                                *request.parallelism, config);
-        }
-
-        pybind11::object tensor =
-            decode_tensor_object_buffer(buffer_ptr, size, error_context);
-        if (tensor.is_none()) {
-            return to_py_ret(ErrorCode::INVALID_PARAMS);
-        }
-        return tensor_route_write(key, tensor, request, config);
     }
 
     int put_manifest_impl(const std::string &key,
@@ -1839,846 +900,13 @@ class MooncakeStorePyWrapper {
         return region;
     }
 
-    std::optional<ParsedTensorMetadata> get_tensor_metadata(
-        const std::string &key,
-        std::shared_ptr<BufferHandle> *buffer_handle_out = nullptr) {
-        if (!is_client_initialized()) {
-            LOG(ERROR) << "Client not initialized";
-            return std::nullopt;
-        }
-
-        std::shared_ptr<BufferHandle> buffer_handle;
-        {
-            py::gil_scoped_release release_gil;
-            buffer_handle = store_->get_buffer(key);
-        }
-        if (!buffer_handle) {
-            return std::nullopt;
-        }
-
-        if (buffer_handle_out) {
-            *buffer_handle_out = buffer_handle;
-        }
-        return ParseTensorMetadata(
-            static_cast<const char *>(buffer_handle->ptr()),
-            buffer_handle->size());
-    }
-
-    std::optional<TensorIntoPlan> build_tensor_into_plan(
-        const std::string &read_key, uintptr_t buffer_ptr, size_t size,
-        const std::string &context,
-        const std::optional<ParsedTensorMetadata> &metadata = std::nullopt) {
-        std::optional<ParsedTensorMetadata> resolved_metadata = metadata;
-        if (!resolved_metadata.has_value()) {
-            resolved_metadata = get_tensor_metadata(read_key);
-        }
-        if (!resolved_metadata.has_value()) {
-            return std::nullopt;
-        }
-
-        const auto total_length =
-            resolved_metadata->data_offset + resolved_metadata->data_bytes;
-        if (total_length > size) {
-            LOG(ERROR) << context << ": buffer too small for key " << read_key;
-            return std::nullopt;
-        }
-
-        auto region =
-            resolve_registered_buffer_region(buffer_ptr, size, context);
-        if (!region.has_value()) {
-            return std::nullopt;
-        }
-        if (region->offset + total_length > region->size) {
-            LOG(ERROR)
-                << context
-                << ": resolved destination range exceeds registered region";
-            return std::nullopt;
-        }
-
-        TensorIntoPlan plan;
-        plan.user_buffer_ptr = buffer_ptr;
-        plan.registered_buffer_ptr = reinterpret_cast<uintptr_t>(region->base);
-        plan.registered_buffer_size = region->size;
-        plan.total_length = total_length;
-        plan.fragments.push_back(TensorIntoFragment{
-            .read_key = read_key,
-            .dst_offset = region->offset,
-            .src_offset = 0,
-            .size = total_length,
-        });
-        return plan;
-    }
-
-    std::optional<TensorIntoPlan> build_full_tensor_into_plan_from_sources(
-        uintptr_t buffer_ptr, size_t size,
-        const std::vector<ReconstructedShardSource> &sources,
-        const std::vector<int64_t> &global_shape, int split_dim, int32_t dtype,
-        const std::string &context, bool allow_empty_fragments = false) {
-        if (sources.empty()) {
-            LOG(ERROR) << context << ": missing reconstruction shard sources";
-            return std::nullopt;
-        }
-
-        auto region =
-            resolve_registered_buffer_region(buffer_ptr, size, context);
-        if (!region.has_value()) {
-            return std::nullopt;
-        }
-
-        size_t representative_index = 0;
-        for (size_t i = 0; i < sources.size(); ++i) {
-            if (sources[i].metadata.data_bytes > 0) {
-                representative_index = i;
-                break;
-            }
-        }
-        const auto &representative = sources[representative_index].metadata;
-        const auto representative_local_shape =
-            TensorShapeToVector(representative.metadata.layout.local_shape,
-                                representative.metadata.header.ndim);
-        if (global_shape.size() != representative_local_shape.size()) {
-            LOG(ERROR) << context << ": invalid tensor metadata shape";
-            return std::nullopt;
-        }
-        if (split_dim < 0 ||
-            split_dim >= static_cast<int>(global_shape.size())) {
-            LOG(ERROR) << context << ": invalid split_dim";
-            return std::nullopt;
-        }
-
-        size_t element_size = 0;
-        if (representative.data_bytes > 0) {
-            int64_t shard_numel = 1;
-            for (auto dim : representative_local_shape) {
-                shard_numel *= dim;
-            }
-            if (shard_numel <= 0 ||
-                representative.data_bytes % static_cast<size_t>(shard_numel) !=
-                    0) {
-                LOG(ERROR) << context << ": invalid shard tensor byte size";
-                return std::nullopt;
-            }
-            element_size =
-                representative.data_bytes / static_cast<size_t>(shard_numel);
-        }
-
-        size_t total_tensor_numel = 1;
-        for (auto dim : global_shape) {
-            total_tensor_numel *= static_cast<size_t>(dim);
-        }
-        const size_t total_tensor_bytes = total_tensor_numel * element_size;
-        const size_t total_length = sizeof(TensorMetadata) + total_tensor_bytes;
-        if (total_length > size ||
-            region->offset + total_length > region->size) {
-            LOG(ERROR) << context
-                       << ": buffer too small for reconstructed tensor";
-            return std::nullopt;
-        }
-
-        TensorMetadata full_metadata = BuildTensorMetadata(
-            dtype, global_shape, global_shape, TensorLayoutKind::FULL);
-        full_metadata.header.data_bytes = total_tensor_bytes;
-        std::memcpy(reinterpret_cast<void *>(buffer_ptr), &full_metadata,
-                    sizeof(TensorMetadata));
-
-        TensorIntoPlan plan;
-        plan.user_buffer_ptr = buffer_ptr;
-        plan.registered_buffer_ptr = reinterpret_cast<uintptr_t>(region->base);
-        plan.registered_buffer_size = region->size;
-        plan.total_length = total_length;
-
-        int64_t elements_before = 1;
-        for (int i = 0; i < split_dim; ++i) {
-            elements_before *= global_shape[i];
-        }
-        int64_t elements_after = 1;
-        for (size_t i = split_dim + 1; i < global_shape.size(); ++i) {
-            elements_after *= global_shape[i];
-        }
-
-        std::vector<size_t> global_strides(global_shape.size(), 1);
-        for (int i = static_cast<int>(global_shape.size()) - 2; i >= 0; --i) {
-            global_strides[i] = global_strides[i + 1] *
-                                static_cast<size_t>(global_shape[i + 1]);
-        }
-
-        int64_t shard_offset = 0;
-        for (size_t i = 0; i < sources.size(); ++i) {
-            const auto &source = sources[i];
-            const auto local_shape =
-                TensorShapeToVector(source.metadata.metadata.layout.local_shape,
-                                    source.metadata.metadata.header.ndim);
-            if (local_shape.size() != global_shape.size()) {
-                LOG(ERROR) << context << ": invalid shard shape for key "
-                           << source.read_key;
-                return std::nullopt;
-            }
-            for (size_t dim = 0; dim < global_shape.size(); ++dim) {
-                if (static_cast<int>(dim) == split_dim) {
-                    continue;
-                }
-                if (local_shape[dim] != global_shape[dim]) {
-                    LOG(ERROR) << context << ": shard shape mismatch for key "
-                               << source.read_key;
-                    return std::nullopt;
-                }
-            }
-
-            const int64_t shard_start = shard_offset;
-            const int64_t shard_size = local_shape[split_dim];
-            shard_offset += shard_size;
-            if (shard_size == 0) {
-                continue;
-            }
-
-            std::vector<size_t> local_strides(local_shape.size(), 1);
-            for (int dim = static_cast<int>(local_shape.size()) - 2; dim >= 0;
-                 --dim) {
-                local_strides[dim] = local_strides[dim + 1] *
-                                     static_cast<size_t>(local_shape[dim + 1]);
-            }
-
-            const size_t chunk_data_offset = source.metadata.data_offset;
-            for (int64_t slice_idx = 0; slice_idx < elements_before;
-                 ++slice_idx) {
-                size_t dst_prefix_linear = 0;
-                size_t src_prefix_linear = 0;
-                int64_t remaining = slice_idx;
-                for (int dim = split_dim - 1; dim >= 0; --dim) {
-                    const int64_t dim_size = global_shape[dim];
-                    const int64_t coord = remaining % dim_size;
-                    remaining /= dim_size;
-                    dst_prefix_linear +=
-                        static_cast<size_t>(coord) * global_strides[dim];
-                    src_prefix_linear +=
-                        static_cast<size_t>(coord) * local_strides[dim];
-                }
-
-                const size_t dst_element_index =
-                    dst_prefix_linear + static_cast<size_t>(shard_start) *
-                                            global_strides[split_dim];
-                const size_t src_element_index = src_prefix_linear;
-                const size_t copy_elements =
-                    static_cast<size_t>(shard_size) *
-                    static_cast<size_t>(elements_after);
-
-                plan.fragments.push_back(TensorIntoFragment{
-                    .read_key = source.read_key,
-                    .dst_offset = region->offset + sizeof(TensorMetadata) +
-                                  dst_element_index * element_size,
-                    .src_offset =
-                        chunk_data_offset + src_element_index * element_size,
-                    .size = copy_elements * element_size,
-                });
-            }
-        }
-
-        if (shard_offset != global_shape[split_dim]) {
-            LOG(ERROR)
-                << context
-                << ": shard extents do not cover reconstructed dimension";
-            return std::nullopt;
-        }
-        if (plan.fragments.empty() &&
-            !(allow_empty_fragments && total_tensor_bytes == 0)) {
-            LOG(ERROR) << context
-                       << ": no fragments planned for reconstruction";
-            return std::nullopt;
-        }
-        return plan;
-    }
-
-    std::optional<TensorIntoPlan> build_tp_full_tensor_into_plan(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const ParallelAxisSpec &axis, const std::string &context,
-        const std::optional<TensorParallelismSpec> &parallelism =
-            std::nullopt) {
-        auto reconstruction = load_tp_full_reconstruction_sources(
-            key, axis, context, parallelism);
-        if (!reconstruction.has_value()) {
-            return std::nullopt;
-        }
-        return build_full_tensor_into_plan_from_sources(
-            buffer_ptr, size, reconstruction->sources,
-            reconstruction->global_shape, reconstruction->split_dim,
-            reconstruction->dtype, context,
-            reconstruction->allow_empty_fragments);
-    }
-
-    std::optional<FullTensorReconstructionSources>
-    load_tp_full_reconstruction_sources(
-        const std::string &key, const ParallelAxisSpec &axis,
-        const std::string &context,
-        const std::optional<TensorParallelismSpec> &parallelism =
-            std::nullopt) {
-        if (axis.size <= 0) {
-            LOG(ERROR) << context << ": tp_size must be positive";
-            return std::nullopt;
-        }
-
-        FullTensorReconstructionSources reconstruction;
-        reconstruction.sources.reserve(axis.size);
-        for (int shard_rank = 0; shard_rank < axis.size; ++shard_rank) {
-            std::string read_key;
-            if (parallelism.has_value()) {
-                auto shard_parallelism = *parallelism;
-                auto tp_axis_index = find_tp_axis_index(shard_parallelism.axes);
-                if (!tp_axis_index.has_value()) {
-                    LOG(ERROR) << context
-                               << ": missing TP axis in full reconstruction";
-                    return std::nullopt;
-                }
-                shard_parallelism.axes[*tp_axis_index].rank = shard_rank;
-                read_key = get_parallelism_key_name(key, shard_parallelism);
-            } else {
-                read_key = resolve_tp_read_key(key, shard_rank, axis.size);
-            }
-            auto metadata = get_tensor_metadata(read_key);
-            if (!metadata.has_value()) {
-                return std::nullopt;
-            }
-            const LayoutAxis *tp_axis =
-                find_layout_axis(metadata->metadata, LayoutAxisKind::TP);
-            if (!is_shard_tensor_metadata(metadata->metadata) || !tp_axis ||
-                tp_axis->shard_rank != shard_rank ||
-                tp_axis->shard_count != axis.size) {
-                LOG(ERROR) << context << ": TP metadata mismatch for key "
-                           << read_key;
-                return std::nullopt;
-            }
-            if (parallelism.has_value()) {
-                auto stored_parallelism =
-                    resolve_tp_compatible_parallelism_from_metadata(
-                        *parallelism, metadata->metadata, context);
-                if (!stored_parallelism.has_value()) {
-                    return std::nullopt;
-                }
-            }
-            reconstruction.sources.push_back(
-                ReconstructedShardSource{read_key, *metadata});
-        }
-
-        reconstruction.global_shape = TensorShapeToVector(
-            reconstruction.sources.front()
-                .metadata.metadata.layout.global_shape,
-            reconstruction.sources.front().metadata.metadata.header.ndim);
-        const LayoutAxis *stored_tp_axis =
-            find_layout_axis(reconstruction.sources.front().metadata.metadata,
-                             LayoutAxisKind::TP);
-        if (!stored_tp_axis) {
-            LOG(ERROR) << context << ": missing TP axis metadata";
-            return std::nullopt;
-        }
-        reconstruction.split_dim = stored_tp_axis->split_dim;
-        if (axis.split_dim.has_value() &&
-            axis.split_dim.value() != reconstruction.split_dim) {
-            LOG(ERROR) << context << ": split_dim mismatch";
-            return std::nullopt;
-        }
-        reconstruction.dtype =
-            reconstruction.sources.front().metadata.metadata.header.dtype;
-        return reconstruction;
-    }
-
-    std::optional<FullTensorReconstructionSources>
-    load_writer_shard_reconstruction(const std::string &key,
-                                     const std::string &context) {
-        if (!is_client_initialized()) {
-            LOG(ERROR) << context << ": client is not initialized";
-            return std::nullopt;
-        }
-
-        std::shared_ptr<BufferHandle> manifest_handle;
-        {
-            py::gil_scoped_release release_gil;
-            manifest_handle =
-                store_->get_buffer(get_writer_manifest_key_name(key));
-        }
-        auto parsed_manifest =
-            parse_writer_shard_manifest(manifest_handle.get());
-        if (!parsed_manifest.has_value()) {
-            return std::nullopt;
-        }
-
-        const auto &manifest = *parsed_manifest;
-        const auto &global_shape = manifest.global_shape;
-        const int split_dim = manifest.manifest.header.split_dim;
-        const int shard_count = manifest.manifest.header.shard_count;
-        if (split_dim < 0 ||
-            split_dim >= static_cast<int>(global_shape.size())) {
-            LOG(ERROR) << context << ": invalid writer split_dim";
-            return std::nullopt;
-        }
-
-        FullTensorReconstructionSources reconstruction;
-        reconstruction.sources.reserve(shard_count);
-        for (int shard_rank = 0; shard_rank < shard_count; ++shard_rank) {
-            WriterPartitionSpec writer{
-                .rank = shard_rank,
-                .size = shard_count,
-                .split_dim = split_dim,
-            };
-            const std::string shard_key =
-                get_writer_shard_key_name(key, writer);
-            auto metadata = get_tensor_metadata(shard_key);
-            if (!metadata.has_value()) {
-                LOG(ERROR) << context << ": missing writer shard key "
-                           << shard_key;
-                return std::nullopt;
-            }
-            auto writer_parallelism =
-                writer_partition_parallelism_from_metadata(metadata->metadata);
-            if (!writer_parallelism.has_value() ||
-                writer_parallelism->axes[0].rank != shard_rank ||
-                writer_parallelism->axes[0].size != shard_count ||
-                writer_parallelism->axes[0].split_dim != split_dim) {
-                LOG(ERROR) << context
-                           << ": writer shard metadata mismatch for key "
-                           << shard_key;
-                return std::nullopt;
-            }
-            reconstruction.sources.push_back(
-                ReconstructedShardSource{shard_key, *metadata});
-        }
-        reconstruction.global_shape = manifest.global_shape;
-        reconstruction.split_dim = manifest.manifest.header.split_dim;
-        reconstruction.dtype = manifest.manifest.header.dtype;
-        reconstruction.allow_empty_fragments = true;
-        return reconstruction;
-    }
-
-    std::optional<TensorIntoPlan> build_writer_shard_full_tensor_into_plan(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const std::string &context) {
-        auto reconstruction = load_writer_shard_reconstruction(key, context);
-        if (!reconstruction.has_value()) {
-            return build_tensor_into_plan(key, buffer_ptr, size, context);
-        }
-
-        return build_full_tensor_into_plan_from_sources(
-            buffer_ptr, size, reconstruction->sources,
-            reconstruction->global_shape, reconstruction->split_dim,
-            reconstruction->dtype, context,
-            reconstruction->allow_empty_fragments);
-    }
-
-    std::optional<FullTensorReconstructionSources>
-    load_parallelism_full_reconstruction_sources(
-        const std::string &key, const TensorParallelismSpec &parallelism,
-        const std::string &context) {
-        const ParallelAxisSpec *request_tp_axis =
-            find_axis_spec_by_kind(parallelism, LayoutAxisKind::TP);
-        if (!request_tp_axis) {
-            LOG(ERROR) << context << ": full reconstruction requires a TP axis";
-            return std::nullopt;
-        }
-
-        auto canonical_parallelism = canonicalize_parallelism_spec(parallelism);
-        if (!canonical_parallelism.has_value()) {
-            LOG(ERROR) << context << ": failed to canonicalize parallelism";
-            return std::nullopt;
-        }
-
-        auto tp_axis_index = find_tp_axis_index(canonical_parallelism->axes);
-        if (!tp_axis_index.has_value()) {
-            LOG(ERROR) << context << ": missing TP axis in request";
-            return std::nullopt;
-        }
-
-        FullTensorReconstructionSources reconstruction;
-        reconstruction.sources.reserve(request_tp_axis->size);
-        for (int shard_rank = 0; shard_rank < request_tp_axis->size;
-             ++shard_rank) {
-            auto shard_parallelism = *canonical_parallelism;
-            shard_parallelism.axes[*tp_axis_index].rank = shard_rank;
-            const std::string shard_key =
-                get_parallelism_key_name(key, shard_parallelism);
-            auto metadata = get_tensor_metadata(shard_key);
-            if (!metadata.has_value()) {
-                LOG(ERROR) << context
-                           << ": no shard matched requested layout for TP rank "
-                           << shard_rank;
-                return std::nullopt;
-            }
-
-            auto stored_parallelism =
-                resolve_tp_compatible_parallelism_from_metadata(
-                    shard_parallelism, metadata->metadata, context);
-            if (!stored_parallelism.has_value() ||
-                !parallelism_specs_equal_by_kind(
-                    *canonical_parallelism, *stored_parallelism,
-                    true /* allow_tp_rank_mismatch */)) {
-                LOG(ERROR) << context
-                           << ": shard metadata mismatch for TP rank "
-                           << shard_rank;
-                return std::nullopt;
-            }
-            reconstruction.sources.push_back(
-                ReconstructedShardSource{shard_key, *metadata});
-        }
-
-        const LayoutAxis *stored_tp_axis =
-            find_layout_axis(reconstruction.sources.front().metadata.metadata,
-                             LayoutAxisKind::TP);
-        if (!stored_tp_axis) {
-            LOG(ERROR) << context << ": missing TP axis metadata";
-            return std::nullopt;
-        }
-        reconstruction.split_dim = stored_tp_axis->split_dim;
-        if (request_tp_axis->split_dim.has_value() &&
-            request_tp_axis->split_dim.value() != reconstruction.split_dim) {
-            LOG(ERROR) << context << ": split_dim mismatch";
-            return std::nullopt;
-        }
-
-        reconstruction.global_shape = TensorShapeToVector(
-            reconstruction.sources.front()
-                .metadata.metadata.layout.global_shape,
-            reconstruction.sources.front().metadata.metadata.header.ndim);
-        reconstruction.dtype =
-            reconstruction.sources.front().metadata.metadata.header.dtype;
-        return reconstruction;
-    }
-
-    std::optional<TensorIntoPlan> build_parallelism_full_tensor_into_plan(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const TensorParallelismSpec &parallelism, const std::string &context) {
-        auto reconstruction = load_parallelism_full_reconstruction_sources(
-            key, parallelism, context);
-        if (!reconstruction.has_value()) {
-            return std::nullopt;
-        }
-        return build_full_tensor_into_plan_from_sources(
-            buffer_ptr, size, reconstruction->sources,
-            reconstruction->global_shape, reconstruction->split_dim,
-            reconstruction->dtype, context,
-            reconstruction->allow_empty_fragments);
-    }
-
-    std::vector<bool> execute_tensor_into_plan_transfers(
-        const std::vector<TensorIntoPlan> &plans) {
-        std::vector<bool> success(plans.size(), false);
-        if (plans.empty()) {
-            return success;
-        }
-
-        std::vector<void *> buffers;
-        std::vector<std::vector<std::string>> all_keys;
-        std::vector<std::vector<std::vector<size_t>>> all_dst_offsets;
-        std::vector<std::vector<std::vector<size_t>>> all_src_offsets;
-        std::vector<std::vector<std::vector<size_t>>> all_sizes;
-        buffers.reserve(plans.size());
-        all_keys.reserve(plans.size());
-        all_dst_offsets.reserve(plans.size());
-        all_src_offsets.reserve(plans.size());
-        all_sizes.reserve(plans.size());
-
-        for (const auto &plan : plans) {
-            buffers.push_back(
-                reinterpret_cast<void *>(plan.registered_buffer_ptr));
-
-            std::unordered_map<std::string, size_t> key_to_index;
-            std::vector<std::string> keys;
-            std::vector<std::vector<size_t>> dst_offsets;
-            std::vector<std::vector<size_t>> src_offsets;
-            std::vector<std::vector<size_t>> sizes;
-
-            for (const auto &fragment : plan.fragments) {
-                if (fragment.read_key.empty() || fragment.size == 0) {
-                    continue;
-                }
-                auto [it, inserted] =
-                    key_to_index.emplace(fragment.read_key, keys.size());
-                if (inserted) {
-                    keys.push_back(fragment.read_key);
-                    dst_offsets.push_back({});
-                    src_offsets.push_back({});
-                    sizes.push_back({});
-                }
-                const size_t key_index = it->second;
-                dst_offsets[key_index].push_back(fragment.dst_offset);
-                src_offsets[key_index].push_back(fragment.src_offset);
-                sizes[key_index].push_back(fragment.size);
-            }
-
-            all_keys.push_back(std::move(keys));
-            all_dst_offsets.push_back(std::move(dst_offsets));
-            all_src_offsets.push_back(std::move(src_offsets));
-            all_sizes.push_back(std::move(sizes));
-        }
-
-        std::vector<std::vector<std::vector<int64_t>>> range_results;
-        {
-            py::gil_scoped_release release_gil;
-            range_results = store_->get_into_ranges(
-                buffers, all_keys, all_dst_offsets, all_src_offsets, all_sizes);
-        }
-
-        for (size_t i = 0; i < plans.size(); ++i) {
-            if (i >= range_results.size() ||
-                range_results[i].size() != all_sizes[i].size()) {
-                continue;
-            }
-
-            success[i] = true;
-            for (size_t key_idx = 0;
-                 key_idx < all_sizes[i].size() && success[i]; ++key_idx) {
-                if (range_results[i][key_idx].size() !=
-                    all_sizes[i][key_idx].size()) {
-                    success[i] = false;
-                    break;
-                }
-                for (size_t frag_idx = 0;
-                     frag_idx < all_sizes[i][key_idx].size(); ++frag_idx) {
-                    if (range_results[i][key_idx][frag_idx] !=
-                        static_cast<int64_t>(all_sizes[i][key_idx][frag_idx])) {
-                        success[i] = false;
-                        break;
-                    }
-                }
-            }
-        }
-        return success;
-    }
-
-    py::list execute_tensor_into_plans(
-        const std::vector<TensorIntoPlan> &plans) {
-        py::list results;
-        for (size_t i = 0; i < plans.size(); ++i) {
-            results.append(py::none());
-        }
-        auto success = execute_tensor_into_plan_transfers(plans);
-        for (size_t i = 0; i < plans.size(); ++i) {
-            if (!success[i]) {
-                continue;
-            }
-            results[i] = buffer_to_tensor(
-                NULL, reinterpret_cast<char *>(plans[i].user_buffer_ptr),
-                static_cast<int64_t>(plans[i].total_length));
-        }
-        return results;
-    }
-
-    std::string resolve_tp_read_key(const std::string &key, int tp_rank,
-                                    int tp_size) const {
-        if (tp_size <= 1) return key;
-        return get_tp_key_name(key, tp_rank);
-    }
-
-    std::vector<std::string> resolve_tp_read_keys(
-        const std::vector<std::string> &base_keys, int tp_rank,
-        int tp_size) const {
-        if (tp_size <= 1) return base_keys;
-
-        std::vector<std::string> shard_keys;
-        shard_keys.reserve(base_keys.size());
-        for (const auto &key : base_keys) {
-            shard_keys.push_back(get_tp_key_name(key, tp_rank));
-        }
-        return shard_keys;
-    }
-
-    pybind11::object get_tensor_with_parallelism(
-        const std::string &key, const py::object &target = py::none()) {
-        auto parsed_target = parse_read_target_spec(target);
-        if (!parsed_target.has_value()) {
-            return pybind11::none();
-        }
-        if (parsed_target->mode == ReadTargetMode::AS_STORED &&
-            !parsed_target->parallelism.has_value()) {
-            return get_tensor(key);
-        }
-        if (parsed_target->mode == ReadTargetMode::FULL) {
-            if (!parsed_target->parallelism.has_value()) {
-                return get_tensor_with_writer_shard_full(
-                    key, "get_tensor_with_parallelism");
-            }
-            auto parallelism = validate_parallelism_spec(
-                parsed_target->parallelism, "ReadTarget(mode=full)", false);
-            if (!parallelism.has_value()) {
-                return pybind11::none();
-            }
-            auto tp_axis_index = find_tp_axis_index(parallelism->axes);
-            if (!tp_axis_index.has_value()) {
-                LOG(ERROR) << "ReadTarget(mode=full): full reconstruction "
-                              "requires a TP axis";
-                return pybind11::none();
-            }
-            const auto &axis = parallelism->axes[*tp_axis_index];
-            return get_tensor_with_tp_full(
-                key, axis.rank, axis.size, axis.split_dim.value_or(0),
-                "get_tensor_with_parallelism", *parallelism);
-        }
-        if (parsed_target->mode == ReadTargetMode::SHARD) {
-            auto parallelism = validate_parallelism_spec(
-                parsed_target->parallelism, "ReadTarget(mode=shard)", false);
-            if (!parallelism.has_value()) {
-                return pybind11::none();
-            }
-            auto resolved = resolve_parallelism_shard_read(
-                [this](const std::string &read_key,
-                       std::shared_ptr<BufferHandle> *buffer_handle_out) {
-                    return get_tensor_metadata(read_key, buffer_handle_out);
-                },
-                key, *parallelism);
-            if (!resolved.has_value()) {
-                LOG(ERROR) << "Parallelism metadata mismatch for key " << key;
-                return pybind11::none();
-            }
-            return buffer_to_tensor(resolved->buffer_handle.get(), NULL, 0);
-        }
-        LOG(ERROR) << "Unsupported ReadTarget mode";
-        return pybind11::none();
-    }
-
-    pybind11::list batch_get_tensor_with_parallelism(
-        const std::vector<std::string> &keys,
-        const py::object &targets = py::none()) {
-        if (targets.is_none()) {
-            return batch_get_tensor(keys);
-        }
-        if (!py::isinstance<py::list>(targets)) {
-            LOG(ERROR) << "targets must be a list or None";
-            py::list empty;
-            for (size_t i = 0; i < keys.size(); ++i) empty.append(py::none());
-            return empty;
-        }
-
-        py::list target_list = py::cast<py::list>(targets);
-        if (target_list.size() != keys.size()) {
-            LOG(ERROR) << "keys and targets must have the same length";
-            py::list empty;
-            for (size_t i = 0; i < keys.size(); ++i) empty.append(py::none());
-            return empty;
-        }
-
-        py::list results;
-        for (size_t i = 0; i < keys.size(); ++i) {
-            results.append(
-                get_tensor_with_parallelism(keys[i], target_list[i]));
-        }
-        return results;
-    }
-
-    std::optional<TensorIntoPlan> build_tensor_into_plan_for_target(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const py::object &target, const std::string &context) {
-        auto parsed_target = parse_read_target_spec(target);
-        if (!parsed_target.has_value()) {
-            return std::nullopt;
-        }
-        if (parsed_target->mode == ReadTargetMode::AS_STORED &&
-            !parsed_target->parallelism.has_value()) {
-            return build_tensor_into_plan(key, buffer_ptr, size, context);
-        }
-
-        if (parsed_target->mode == ReadTargetMode::FULL &&
-            !parsed_target->parallelism.has_value()) {
-            return build_writer_shard_full_tensor_into_plan(key, buffer_ptr,
-                                                            size, context);
-        }
-
-        auto parallelism = validate_parallelism_spec(parsed_target->parallelism,
-                                                     context, false);
-        if (!parallelism.has_value()) {
-            return std::nullopt;
-        }
-
-        if (parsed_target->mode == ReadTargetMode::SHARD) {
-            auto resolved = resolve_parallelism_shard_read(
-                [this](const std::string &read_key,
-                       std::shared_ptr<BufferHandle> *buffer_handle_out) {
-                    return get_tensor_metadata(read_key, buffer_handle_out);
-                },
-                key, *parallelism);
-            if (!resolved.has_value()) {
-                LOG(ERROR) << context
-                           << ": parallelism metadata mismatch for key " << key;
-                return std::nullopt;
-            }
-            return build_tensor_into_plan(resolved->read_key, buffer_ptr, size,
-                                          context, resolved->metadata);
-        }
-
-        if (parsed_target->mode == ReadTargetMode::FULL) {
-            return build_parallelism_full_tensor_into_plan(
-                key, buffer_ptr, size, *parallelism, context);
-        }
-
-        LOG(ERROR) << context << ": unsupported ReadTarget mode";
-        return std::nullopt;
-    }
-
-    pybind11::object get_tensor_with_parallelism_into(
-        const std::string &key, uintptr_t buffer_ptr, size_t size,
-        const py::object &target = py::none()) {
-        auto plan = build_tensor_into_plan_for_target(
-            key, buffer_ptr, size, target, "get_tensor_with_parallelism_into");
-        if (!plan.has_value()) {
-            return pybind11::none();
-        }
-        py::list results = execute_tensor_into_plans({*plan});
-        if (results.empty()) {
-            return py::none();
-        }
-        return py::reinterpret_borrow<py::object>(results[0]);
-    }
-
-    pybind11::list batch_get_tensor_with_parallelism_into(
-        const std::vector<std::string> &keys,
-        const std::vector<uintptr_t> &buffer_ptrs,
-        const std::vector<size_t> &sizes,
-        const py::object &targets = py::none()) {
-        if (targets.is_none()) {
-            return batch_get_tensor_into(keys, buffer_ptrs, sizes);
-        }
-        if (!py::isinstance<py::list>(targets)) {
-            LOG(ERROR) << "targets must be a list or None";
-            py::list empty;
-            for (size_t i = 0; i < keys.size(); ++i) empty.append(py::none());
-            return empty;
-        }
-
-        py::list target_list = py::cast<py::list>(targets);
-        if (target_list.size() != keys.size() ||
-            buffer_ptrs.size() != keys.size() || sizes.size() != keys.size()) {
-            LOG(ERROR) << "keys, buffer_ptrs, sizes, and targets must have the "
-                          "same length";
-            py::list empty;
-            for (size_t i = 0; i < keys.size(); ++i) empty.append(py::none());
-            return empty;
-        }
-
-        std::vector<TensorIntoPlan> valid_plans;
-        std::vector<size_t> valid_indices;
-        valid_plans.reserve(keys.size());
-        valid_indices.reserve(keys.size());
-        py::list results;
-        for (size_t i = 0; i < keys.size(); ++i) {
-            results.append(py::none());
-            auto plan = build_tensor_into_plan_for_target(
-                keys[i], buffer_ptrs[i], sizes[i], target_list[i],
-                "batch_get_tensor_with_parallelism_into");
-            if (!plan.has_value()) {
-                continue;
-            }
-            valid_indices.push_back(i);
-            valid_plans.push_back(*plan);
-        }
-
-        py::list valid_results = execute_tensor_into_plans(valid_plans);
-        for (size_t i = 0; i < valid_indices.size() && i < valid_results.size();
-             ++i) {
-            results[valid_indices[i]] = valid_results[i];
-        }
-        return results;
-    }
-
+#include "store_py_parallel_read.h"
 
     // --- Upsert tensor methods ---
 
-    int upsert_tensor_info_impl(const std::string &key, const PyTensorInfo &info,
-                           const ReplicateConfig &config) {
+    int upsert_tensor_info_impl(const std::string &key,
+                                const PyTensorInfo &info,
+                                const ReplicateConfig &config) {
         if (!info.valid()) return to_py_ret(ErrorCode::INVALID_PARAMS);
 
         std::vector<std::span<const char>> values;
@@ -2751,19 +979,6 @@ class MooncakeStorePyWrapper {
         }
     }
 
-    int upsert_tensor_parallelism_tp_impl(
-        const std::string &key, pybind11::object tensor,
-        const ReplicateConfig &config, int tp_size, int split_dim,
-        const std::vector<ParallelAxisSpec> &axes) {
-        return execute_tp_tensor_write_impl(
-            key, tensor, config, tp_size, split_dim, axes,
-            "Failed to upsert tensor with tp",
-            [this](const std::string &shard_key, const PyTensorInfo &info,
-                   const ReplicateConfig &write_config) {
-                return upsert_tensor_info_impl(shard_key, info, write_config);
-            });
-    }
-
     std::vector<int> batch_upsert_tensor_from(
         const std::vector<std::string> &keys,
         const std::vector<uintptr_t> &buffer_ptrs,
@@ -2811,7 +1026,6 @@ class MooncakeStorePyWrapper {
         return store_->batch_upsert_from(keys, buffers, sizes,
                                          ReplicateConfig{});
     }
-
 
     std::vector<int> batch_upsert_tensor_impl(
         const std::vector<std::string> &keys,
